AI-based/requirements.txt
django==3.2
djangorestframework==3.13.1
pillow==9.0.0
